Initial commit for the project.
A basic website with a CRUD template in code, using HTMX to provide PUT
and DELETE on forms, buttons, etc, and to replace content on the page
with what the server is sending.

The website is backed up an sqlite database with a basic schema
migration system set up.

The HTML structure for data and its update forms are driven by
typeclasses, allowing for the form to be defined once, and wrapper
functions to make building a page easy.