Documenting code, and some code fixes.
Adding documentation to a lot of the code.
Fixing schema migrations. The migration code was broken as there was
never an initial version set in the schema_version table. It now has one
set as part of an initial migration that will happen after that tabel is
made.
Various small code cleanups that I noticed while writing docs.