Adding authentication code for users.
Expanding the users code to include more database tables and lookups.
This is a new schema and set of types and functions. These will need to
have tests written for them to ensure that they work as expected.

Filling out the authentication code to check a user against the
database. Passwords are hashed with Argon2, and stored in the local
database in a seperate table. This is a thin seperation between
particularly sensitive data and the user record that will be used in
most of the application code, but it won't hurt to have it there anyway.

Started a Hedgehog state machine to eventually test the API. Since this
is a HTML driving API, for the moment it will focus on API
authentication and authorisation, rather than checking that the response
HTML is structured in a particular way.