Expanding state machine testing for the server.
Reworked the test routes to work around an issue where a `Raw`
combinator was trapping all of the requests, essentially blackholing our
testing requests. Moving the test routes to be first in the API type
fixed the issue.

Split up the main server code to allow passing in custom servers and API
types, as this means most of the code can be shared between the main
server and the testing server. The main difference now is the addition
of the testing routes to the API.