Changing testing state to be map based, and using Lens
Changing the main testing state to be Map based so that building tests
can be a bit smarter about generating values, and we can use Map
functions for updates and deletions.

Adding Lens to the tests to help keep names under control. Using
`HasField` was working for a while, but some GHC warnings about
ambiguous updates and deprecated behaviour made me want to move off that
pattern an onto something I'm more familiar with.