Adding SSL and more monad reworking.
Adding basic SSL certificate support.
Reworking the monad stack so that auth isn't passed around directly, but
rather so that auth is verified, users are looked up (and possibly a
unauth is thrown if they no longer exist) and then their display info is
stored in a local env for further use.