CSS, route, and content changes.
Small changes to the CSS to make the site slightly nicer.
Changes to the routes and their protection to allow for unauthentricated
users to list and read entries.
Adding Commonmark support for entry content. This is rendered to HTML
server side, with suspicious HTML being filtered out by the cmark
library.
Adding more bits of text around the site to better explain what is
happening.
Adding recent entries to a list on the homepage.